+++ android/src/test/java/com/example/ivs_broadcaster/IvsBroadcasterPluginTest.java
@@ -22,7 +22,7 @@ public class IvsBroadcasterPluginTest {
 
     final MethodCall call = new MethodCall("getPlatformVersion", null);
     MethodChannel.Result mockResult = mock(MethodChannel.Result.class);
-    plugin.onMethodCall(call, mockResult);
+
 
     verify(mockResult).success("Android " + android.os.Build.VERSION.RELEASE);
   }